ControlPanel/requirements.txt
Flask==3.1.0
Flask-SocketIO==5.4.1
paho-mqtt==2.1.0
plotly==5.24.1
eventlet==0.38.0
gunicorn==23.0.0
gevent
flask_mqtt
gevent-websocket
pymysql
flask_sqlalchemy